why tf does the page just kill itself at the end instead of scrolling
