+++ frontend/src/components/css/Gallery.scss
@@ -1,3 +1,12 @@
+.gallery-mentor-container {
+  padding: 1vh;
+  display: grid;
+  grid-template-columns: repeat(auto-fill, minmax(30%, 1fr));
+  grid-gap: 1rem;
+  height: 100%;
+  overflow-y: scroll;
+}
+
 .gallery-mentor-card {
     background: #FFFFFF;
     border: 2px solid #E4BB4F;
@@ -6,15 +15,6 @@
 
 .gallery-mentor-card:hover {
     box-shadow: 0px 4px 4px rgba(0, 0, 0, 0.25);
-}
-
-.gallery-mentor-container {
-    padding: 1vh;
-    display: grid;
-    grid-template-columns: repeat(auto-fill, minmax(30%, 1fr));
-    grid-gap: 1rem;
-    height: 100%;
-    overflow-y: scroll;
 }
 
 .gallery-lesson-types {
@@ -32,7 +32,7 @@
   display: inline-block;
 }
 
-hr.solid  {
+.gallery-solid-border  {
   border-top: 3px solid #F2C94C;
   margin-top: 20px;
 }
@@ -40,9 +40,11 @@ hr.solid  {
 .gallery-profile-button {
     float: right;
     background: #E4BB4F;
+    border-color: #E4BB4F;
     color: white;
     margin-bottom: 15px;
     margin-top: 10px;
+    border-radius: 13px;
 }
 
 .gallery-profile-button:hover, .gallery-profile-button:focus {
