+++ frontend/src/components/pages/Gallery.js
@@ -1,15 +1,8 @@
 import React, { useState, useEffect } from "react";
-import {
-  LinkOutlined,
-  LinkedinOutlined,
-  StarOutlined,
-  EnvironmentOutlined,
-  UserOutlined,
-} from "@ant-design/icons";
-import "../css/Gallery.scss";
 import { fetchMentors } from "../../utils/api";
-import { Avatar, Typography, Button } from "antd";
-const { Title, Text } = Typography;
+import MentorCard from "../MentorCard";
+
+import "../css/Gallery.scss";
 
 function Gallery() {
   const [mentors, setMentors] = useState([]);
@@ -47,6 +40,7 @@ function Gallery() {
           specializations={mentor.specializations}
           website={mentor.website}
           linkedin={mentor.linkedin}
+          id={mentor._id["$oid"]}
           lesson_types={getLessonTypes(
             mentor.offers_group_appointments,
             mentor.offers_in_person
@@ -56,78 +50,5 @@ function Gallery() {
     </div>
   );
 }
-
-function MentorCard(props) {
-  return (
-    <div className="gallery-mentor-card">
-      <div className="gallery-card-body">
-        <div className="gallery-card-header">
-          <Avatar size={90} icon={getPicture(props.picture)} />
-          <div className="gallery-header-text">
-            <Title style={styles.title} className="gallery-title-text">
-              {props.name}
-            </Title>
-            <Title style={styles.title} type="secondary" level={5}>
-              {props.professional_title}
-            </Title>
-            <Title style={styles.title} type="secondary" level={5}>
-              Speaks: {props.languages.join(", ")}
-            </Title>
-          </div>
-        </div>
-        <h3 className="gallery-lesson-types">
-          <span class="dot" />
-          {props.lesson_types}
-        </h3>
-        <h3 className="gallery-headers">
-          <EnvironmentOutlined style={styles.icon} />
-          Location:
-        </h3>
-        <Text className="gallery-list-items">{props.location}</Text>
-        <h3 className="gallery-headers">
-          <StarOutlined style={styles.icon} />
-          Specializations:
-        </h3>
-        <Text className="gallery-list-items">
-          {props.specializations.join(", ")}
-        </Text>
-        <h4>
-          <LinkOutlined style={styles.icon} />
-          <a className="gallery-links" href={props.website}>
-            {props.website}
-          </a>
-        </h4>
-        <h4>
-          <LinkedinOutlined style={styles.icon} />
-          <a className="gallery-links" href={props.linkedin}>
-            linkedin
-          </a>
-        </h4>
-        <hr class="solid" />
-        <Button className="gallery-profile-button" shape="round" size="medium">
-          View Profile
-        </Button>
-      </div>
-    </div>
-  );
-}
-
-function getPicture(picture) {
-  if (!picture) {
-    return <UserOutlined />;
-  } else {
-    return <img src={picture} alt="" />;
-  }
-}
-
-const styles = {
-  title: {
-    margin: 0,
-  },
-  icon: {
-    fontSize: "20px",
-    "padding-right": "7px",
-  },
-};
 
 export default Gallery;
